added note removal

+++ main.cpp
@@ -10,8 +10,8 @@ int main(int argc, char *argv[])
     constexpr int WIDTH = 1920 * 0.5;
     constexpr int HEIGHT = 1080 * 0.5;
 
-    // Type determines the quality of the audio short < int < float < double
-    auto &app = APPLICATION<double>::instance(WIDTH, HEIGHT);
+    // Type determines the quality of the audio float < double
+    auto &app = APPLICATION<float>::instance(WIDTH, HEIGHT);
     app.run();
     return 0;
 }
